+++ src/connectors/impls/otel/id.rs
@@ -105,8 +105,7 @@ fn hex_id_to_pb(
         json.to_vec()
     } else if let Some(arr) = data.as_array() {
         arr.iter().map(Value::as_u8).collect::<Option<Vec<u8>>>().ok_or_else(|| format!(
-            "Invalid {} id ( wrong array element ) - values must be between 0 and 255 - cannot convert to pb",
-            kind
+            "Invalid {kind} id ( wrong array element ) - values must be between 0 and 255 - cannot convert to pb"
         ))?
     } else {
         return Err(format!("Cannot convert json value to otel pb {kind} id").into());
@@ -114,11 +113,7 @@ fn hex_id_to_pb(
     if (allow_empty && data.is_empty()) || data.len() == len_bytes {
         Ok(data)
     } else {
-        Err(format!(
-            "Invalid {} id ( wrong array length ) - cannot convert to pb",
-            kind
-        )
-        .into())
+        Err(format!("Invalid {kind} id ( wrong array length ) - cannot convert to pb",).into())
     }
 }