+++ CHANGELOG.md
@@ -11,7 +11,7 @@
 
 ### New Features
 
-* Update to rust toolchain 1.66
+* Update to rust toolchain 1.68
 * Warning types: now warnings can specify what they are about (performance, behaviour, etc.)
 * New warnings for consistency, behaviour and performance
 * Functions now can create warnings